Add files via upload
In this project, the two different post types from Hacker News, are analysed.
These are Ask HN and Show HN.

Users Ask HN when they want to post questions they need answers to while the
Submit HN is to reply, answer questions posted or just post an interesting update.

The aim of this project is to determine:
1. Find out which posts receive more comments on average
2. Find out if posts created at a certain time receive more comments on average.

Note that the data explored is reduced from 300,000 rows to about 20,000 by removing all submissions that did not receive any comments.

In conclusion, ask posts and show posts were analyzed to determine which type of post and time
receive the most comments on average. Based on the analysis, to maximize the amount of comments a post receives, I recommend the post be categorized as ask post and created between 15:00 and 16:00.

However, it should be noted that the dataset analyzed excluded posts without any comments. Given that, it's more accurate to say that out of the posts that received comments, ask posts received more comments on average and the ask posts created between 15:00 and 16:00 received the most comments on average.